added Proactor.run_test(); test updates
Proactor.run_test() takes one or more 'test steps' as arugment.
Each test-step is posted in-turn, but only when the proactor is /idle/,
which is defined as not having run a handler in a (small) time interval.

This provides a guarantee that 1) all servers in the test fixture have
fully completed server bootstrap, or that 2) all operations related to
some previous test phase have completed, prior to beginning the next
test phase.